+++ internal/view/widget.go
@@ -3,11 +3,14 @@ package view
 import (
 	"context"
 	"time"
+
+	"github.com/slok/meterm/internal/view/template"
 )
 
 type syncConfig struct {
 	timeRangeStart time.Time
 	timeRangeEnd   time.Time
+	templateData   template.Data
 }
 
 type widget interface {